// src/core/http/generated/models/invoice.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { InvoiceMetadata } from './invoiceMetadata.ts';
import type { InvoiceStatus } from './invoiceStatus.ts';
import type { Subscription } from './subscription.ts';
import type { Tenant } from './tenant.ts';

export interface Invoice {
  id: string;
  number: string;
  tenant_id?: string;
  tenant?: Tenant;
  subscription_id?: string;
  subscription?: Subscription;
  gateway?: string;
  status: InvoiceStatus;
  amount: number;
  currency: string;
  due_at?: string;
  paid_at?: string;
  gateway_invoice_id?: string;
  invoice_url?: string;
  metadata?: InvoiceMetadata;
  created_at: string;
  updated_at?: string;
}